# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "All APIs related to Karchu.",
        "title": "Karchu API",
        "contact": {},
        "version": "1.0"
    },
    "host": "karchu.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/v2/money-lent": {
            "post": {
                "description": "Money lent to a friend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MoneyLent"
                ],
                "summary": "Money Lent to a friend",
                "parameters": [
                    {
                        "description": "enter Email, Password and friend Name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.MoneyLentFriend"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/settle-transaction": {
            "post": {
                "description": "settle a split transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settleTransaction"
                ],
                "summary": "settle a split",
                "parameters": [
                    {
                        "description": "enter Email, Password and friend Name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.SettleTransactionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "settle a split transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settleTransaction"
                ],
                "summary": "settle a split",
                "parameters": [
                    {
                        "description": "enter Email, Password and friend Name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UnSettleTransactionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/settle-transaction/friend": {
            "post": {
                "description": "settle a split of a friend",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settleTransaction"
                ],
                "summary": "settle a split of a friend",
                "parameters": [
                    {
                        "description": "enter Email, Password and friend Name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.SettleTransactionFriend"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/settle-transaction/str": {
            "post": {
                "description": "settle a split transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "settleTransaction"
                ],
                "summary": "settle a split",
                "parameters": [
                    {
                        "description": "settle transaction",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.SettleTransactionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/split-transaction": {
            "post": {
                "description": "split a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SplitTransaction"
                ],
                "summary": "split a transaction",
                "parameters": [
                    {
                        "description": "split transaction",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.SplitTransactionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes a alreadt split transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SplitTransaction"
                ],
                "summary": "deletes a alreadt split transaction",
                "parameters": [
                    {
                        "description": "delete split",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.DeleteSplitTransactionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/split-transaction/one": {
            "post": {
                "description": "split a transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SplitTransaction"
                ],
                "summary": "split a transaction",
                "parameters": [
                    {
                        "description": "split with one",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.SplitWithOneFriendReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/split-transaction/settled-splits": {
            "post": {
                "description": "Get splits of user which are settled",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SplitTransaction"
                ],
                "summary": "Get splits of user which are settled",
                "parameters": [
                    {
                        "description": "get split transaction",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.GetSettledSplitTransactionsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/split-transaction/splits": {
            "post": {
                "description": "Get splits of user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SplitTransaction"
                ],
                "summary": "Get splits of user",
                "parameters": [
                    {
                        "description": "get split transaction",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.GetSplitTransactionsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/split-transaction/str": {
            "delete": {
                "description": "deletes a alreadt split transaction from transString",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SplitTransaction"
                ],
                "summary": "deletes a alreadt split transaction from transString",
                "parameters": [
                    {
                        "description": "delete split",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.DeleteSplitTransactionStringReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
        "/v2/split-transaction/unsettled-splits": {
            "post": {
                "description": "Get splits of user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SplitTransaction"
                ],
                "summary": "Get splits of user",
                "parameters": [
                    {
                        "description": "get split transaction",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.GetUnSettledSplitTransactionsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "requests.DeleteSplitTransactionReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "integer"
                }
            }
        },
        "requests.DeleteSplitTransactionStringReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "transString": {
                    "type": "string"
                }
            }
        },
...
        "requests.FriendSplit": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "friendId": {
                    "type": "integer"
                }
            }
        },
...
        "requests.GetSettledSplitTransactionsReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "requests.GetSplitTransactionsReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.GetUnSettledSplitTransactionsReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.MoneyLentFriend": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "friendName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "requests.SettleTransactionFriend": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "friendName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "requests.SettleTransactionReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "splitTransactionId": {
                    "type": "integer"
                }
            }
        },
        "requests.SplitTransactionReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "splits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/requests.FriendSplit"
                    }
                },
                "transactionId": {
                    "type": "integer"
                }
            }
        },
        "requests.SplitWithOneFriendReq": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "friend": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "transString": {
                    "type": "string"
                }
            }
        },
...
        "requests.UnSettleTransactionReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "splitTransactionId": {
                    "type": "integer"
                }
            }
        },
...
        "responses.SuccessRes": {
            "type": "object",
            "properties": {
                "success_code": {
                    "type": "string"
                },
                "success_message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}